+++ cmd/root.go
@@ -1,19 +1,14 @@
 package cmd
 
 import (
-	"io"
 	"log"
 	"os"
 	"strings"
 
 	"github.com/spf13/cobra"
 
-	"github.com/giantswarm/microerror"
-	"github.com/giantswarm/micrologger"
-
 	"github.com/giantswarm/architect/cmd/create"
 	"github.com/giantswarm/architect/cmd/helm"
-	"github.com/giantswarm/architect/cmd/legacy"
 	"github.com/giantswarm/architect/cmd/preparerelease"
 	cmdProject "github.com/giantswarm/architect/cmd/project"
 )
@@ -35,37 +30,6 @@ var (
 
 func init() {
 	var err error
-
-	var logger micrologger.Logger
-	{
-		c := micrologger.Config{}
-
-		logger, err = micrologger.New(c)
-		if err != nil {
-			panic(microerror.Pretty(microerror.Mask(err), true))
-		}
-
-	}
-
-	var stderr, stdout io.Writer
-	{
-		stderr = os.Stderr
-		stdout = os.Stdout
-	}
-
-	var legacyCmd *cobra.Command
-	{
-		c := legacy.Config{
-			Logger: logger,
-			Stderr: stderr,
-			Stdout: stdout,
-		}
-
-		legacyCmd, err = legacy.New(c)
-		if err != nil {
-			panic(microerror.Pretty(microerror.Mask(err), true))
-		}
-	}
 
 	// Get the current working directory, to use as a default.
 	defaultWorkingDirectory, err := os.Getwd()
@@ -103,6 +67,5 @@ func init() {
 	RootCmd.AddCommand(cmdProject.Cmd)
 	RootCmd.AddCommand(create.Cmd)
 	RootCmd.AddCommand(helm.Cmd)
-	RootCmd.AddCommand(legacyCmd)
 	RootCmd.AddCommand(preparerelease.Cmd)
 }